third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/registries/ShadowRoot-init-declarative-expected.txt

This is a testharness.js-based test.
[FAIL] Custom element inside 'shadowrootcustomelementregistry' declarative shadow root should use document's registry as attachShadow default registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Built-in element inside 'shadowrootcustomelementregistry' declarative shadow root should use document's registry as attachShadow default registry
  assert_equals: expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
